requirements/requirements.txt
django>=1.9,<1.10
django-admin-bootstrapped
django-bootstrap3==7.0.1
django-bower==5.1.0
django-braces==1.8.1
django-compressor==2.0
django-crispy-forms==1.6.0
django-exclusivebooleanfield
django-extensions==1.6.1
django-extra-views==0.7.1
django-filter==0.13.0
django-floppyforms==1.6.1
django-image-cropping
django-model-utils==2.4
django-modelcluster==2.0
django-sass-processor==0.3.4
django-smart-selects
django-taggit==0.21.2
djangorestframework==3.9.1
dj-database-url==0.4.1
easy-thumbnails==2.3
trml2pdf
gunicorn==19.5.0
libsass==0.11.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyjwkest
python-dateutil
python-decouple==3.0
python-magic==0.4.10
pytz==2016.3
pyyaml==5.1
rtyaml==0.0.2
sqlparse==0.2.0
unipath==1.1
xlwt
